Fix Kotlin compilation: move issueCentralToken outside synchronized block and cast distanceKm to Double

+++ app/src/main/java/com/example/shribalajikripadham/data/network/HostingerCentralSyncManager.kt
@@ -256,7 +256,7 @@ object HostingerCentralSyncManager {
                 deviceId = token.deviceId,
                 latitude = token.latitude,
                 longitude = token.longitude,
-                distanceKm = token.distanceKm,
+                distanceKm = token.distanceKm.toDouble(),
                 photoUrl = token.photoUri,
                 registeredBy = token.registeredBy,
                 originAddress = token.originAddress,
